refactor: Move prompt logic to internal/architect/prompt package
Move the StitchPrompt and EscapeContent functions from output.go
to the new prompt package. Update all references to use these
functions from their new location. This improves separation of
concerns by putting prompt-related functionality in a dedicated package.
